# docs/swagger.yaml
basePath: /swagger/ebcp-service
definitions:
...
  common.PageGeneric-model_Ebcp_control_device:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Ebcp_control_device'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
...
  common.Response:
    properties:
      data: {}
      msg:
        type: string
      status:
        type: integer
    type: object
  model.Ebcp_control_device:
    properties:
      commands:
        description: commands
        type: string
      created_by:
        description: created_by
        type: string
      created_time:
        description: created_time
        type: string
      device_type:
        description: 设备类型
        type: string
      id:
        description: id
        type: string
      ip_address:
        description: ip_address
        type: string
      name:
        description: 设备名称
        type: string
      port:
        description: port
        type: integer
      room_id:
        description: 所属展厅ID
        type: string
      status:
        description: '状态(1: 正常, 2: 故障)'
        type: integer
      updated_by:
        description: updated_by
        type: string
      updated_time:
        description: updated_time
        type: string
      version:
        description: version
        type: string
    type: object
...
info:
  contact: {}
  description: ebcp-service API
  title: ebcp-service API
  version: "1.0"
paths:
...
  /ebcp-control-device:
    get:
      description: query objects
      parameters:
      - description: _select
        in: query
        name: _select
        type: string
      - description: order
        in: query
        name: _order
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: device_type
        in: query
        name: device_type
        type: string
      - description: ip_address
        in: query
        name: ip_address
        type: string
      - description: port
        in: query
        name: port
        type: string
      - description: version
        in: query
        name: version
        type: string
      - description: room_id
        in: query
        name: room_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: commands
        in: query
        name: commands
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Ebcp_control_device'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: query objects
      tags:
      - 中控设备
    post:
      consumes:
      - application/json
      description: save
      parameters:
      - description: object
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/model.Ebcp_control_device'
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ebcp_control_device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: save
      tags:
      - 中控设备
  /ebcp-control-device/{id}:
    delete:
      description: delete
      parameters:
      - description: 实例id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: object
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Ebcp_control_device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: delete
      tags:
      - 中控设备
  /ebcp-control-device/batch-delete:
    post:
      consumes:
      - application/json
      description: batch delete
      parameters:
      - description: id array
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch delete
      tags:
      - 中控设备
  /ebcp-control-device/batch-upsert:
    post:
      consumes:
      - application/json
      description: batch update
      parameters:
      - description: objects array
        in: body
        name: entities
        required: true
        schema:
          items:
            additionalProperties: true
            type: object
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: batch update
      tags:
      - 中控设备
  /ebcp-control-device/page:
    get:
      description: page query, _page(from 1 begin), _page_size, _order, and others
        fields, status=1, name=$like.%CAM%
      parameters:
      - description: current page
        in: query
        name: _page
        required: true
        type: integer
      - description: page size
        in: query
        name: _page_size
        required: true
        type: integer
      - description: order
        in: query
        name: _order
        type: string
      - description: _select
        in: query
        name: _select
        required: true
        type: string
      - description: id
        in: query
        name: id
        type: string
      - description: created_by
        in: query
        name: created_by
        type: string
      - description: created_time
        in: query
        name: created_time
        type: string
      - description: updated_by
        in: query
        name: updated_by
        type: string
      - description: updated_time
        in: query
        name: updated_time
        type: string
      - description: name
        in: query
        name: name
        type: string
      - description: device_type
        in: query
        name: device_type
        type: string
      - description: ip_address
        in: query
        name: ip_address
        type: string
      - description: port
        in: query
        name: port
        type: string
      - description: version
        in: query
        name: version
        type: string
      - description: room_id
        in: query
        name: room_id
        type: string
      - description: status
        in: query
        name: status
        type: string
      - description: commands
        in: query
        name: commands
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: objects array
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/common.PageGeneric-model_Ebcp_control_device'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.Response'
      summary: page query
      tags:
      - 中控设备
...
swagger: "2.0"